% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{bvals_to_mvals}
\alias{bvals_to_mvals}
\title{Convert beta values to M-values}
\usage{
bvals_to_mvals(bvals)
}
\arguments{
\item{bvals}{Numeric vector or matrix of beta values (between 0 and 1)}
}
\value{
Numeric vector or matrix of M-values
}
\description{
This function converts beta values to M-values.
}
\examples{
bvals <- c(0.1, 0.3, 0.5, 0.7, 0.9)
mvals <- bvals_to_mvals(bvals)
}
