% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{F1_logli}
\alias{F1_logli}
\title{Log-linear transformation function for Clock 3}
\usage{
F1_logli(age1, m1, m2 = m1, c1 = 1)
}
\arguments{
\item{age1}{Numeric vector of ages to transform}

\item{m1}{Numeric parameter m1 for transformation}

\item{m2}{Numeric parameter m2 for transformation (default: same as m1)}

\item{c1}{Numeric parameter c1 for transformation (default: 1)}
}
\value{
Numeric vector of transformed values
}
\description{
This function performs the log-linear transformation used in Universal Clock 3.
}
