% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{F2_revtrsf}
\alias{F2_revtrsf}
\title{Reverse transformation function for Clock 3}
\usage{
F2_revtrsf(y.pred, m1, m2 = m1, c1 = 1)
}
\arguments{
\item{y.pred}{Numeric vector of predicted values to reverse transform}

\item{m1}{Numeric parameter m1 for transformation}

\item{m2}{Numeric parameter m2 for transformation (default: same as m1)}

\item{c1}{Numeric parameter c1 for transformation (default: 1)}
}
\value{
Numeric vector of original scale values
}
\description{
This function performs the reverse transformation for Universal Clock 3 predictions.
}
