% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_prediction_functions.R
\name{get_clock_info}
\alias{get_clock_info}
\title{Get Clock Information}
\usage{
get_clock_info(clock_type = NULL, verbose = TRUE)
}
\arguments{
\item{clock_type}{Character string to filter by clock type (optional)}

\item{verbose}{Logical indicating whether to print detailed information (default: TRUE)}
}
\value{
Data frame with clock information
}
\description{
Get information about available clocks in the package.
}
\examples{
\dontrun{
# Get all clock information
clock_info <- get_clock_info()

# Get only ensemble static clocks
ensemble_static <- get_clock_info("EnsembleAge.Static")
}
}
