% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient_data_loading.R
\name{get_required_clock_probes}
\alias{get_required_clock_probes}
\title{Get required CpG probes for all clocks}
\usage{
get_required_clock_probes(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical indicating whether to print progress messages (default: TRUE)}
}
\value{
Character vector of required CpG IDs
}
\description{
This function extracts all CpG probes required by any of the clocks
to enable efficient data loading.
}
\examples{
\dontrun{
required_probes <- get_required_clock_probes()
}
}
