% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{mvals_to_bvals}
\alias{mvals_to_bvals}
\title{Convert M-values to beta values}
\usage{
mvals_to_bvals(mvals)
}
\arguments{
\item{mvals}{Numeric vector or matrix of M-values}
}
\value{
Numeric vector or matrix of beta values
}
\description{
This function converts M-values (log2 ratio of methylated to unmethylated intensities)
to beta values (proportion of methylated intensities).
}
\examples{
mvals <- c(-2, -1, 0, 1, 2)
bvals <- mvals_to_bvals(mvals)
}
